add: run5 experiment
This experiment is updated to show job and experiment completion
times on the plot (the temporal cost of waiting for a scale down)
and adding flux job times to the pod times plot, which is not
a totally fair comparison but the best data I have at the moment.
We also extend the plot to show the nodes going down, and update
the text.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>